refactor(tests): share the synthetic run skeleton builders via conftest (#25)
The run report tests need the same in-memory skeleton and evaluation builders
solutions_test already had. Moving them to conftest keeps one copy rather than
growing a second.
